refactor: move join/component to /component
